+++ bfm/yala/src/camera.cpp
@@ -6,7 +6,7 @@ Camera::Camera(size_t screenWidth, size_t screenHeight)
 {
   displayResolutionX = screenWidth;
 
-  aspectRatio = (float)screenWidth / (float)screenHeight;
+  aspectRatio = (double)screenWidth / (double)screenHeight;
   fov = 45.0f;
   nearClip = 1.0f;
   farClip = 1000.0f;
@@ -18,35 +18,36 @@ Camera::Camera(size_t screenWidth, size_t screenHeight)
   roll = 0;
   speed = 5;
 
-  movement = glm::vec3(0);
-  position = glm::vec3(0);
+  movement = glm::dvec3(0);
+  position = glm::dvec3(0);
 
-  up = glm::vec3(0,1,0);
-  forward = glm::vec3(0,0,-1);
-  right = glm::vec3(1,0,0);
+  up = glm::dvec3(0,1,0);
+  forward = glm::dvec3(0,0,-1);
+  right = glm::dvec3(1,0,0);
 
-  viewMatrix = glm::mat4(1);
-  projectionMatrix = glm::mat4(1);
-  viewProjectionMatrix = glm::mat4(1);
+  viewMatrix = glm::dmat4(1);
+  projectionMatrix = glm::dmat4(1);
+  viewProjectionMatrix = glm::dmat4(1);
 }
 
-void Camera::Update(float elapsedMS)
+void Camera::Update(double elapsedMS)
 {
   // Determine where the local XYZ axes are pointing...
-  const glm::mat4 orientation = glm::yawPitchRoll(yaw, pitch, roll);
+  const glm::dmat4 orientation = glm::yawPitchRoll(yaw, pitch, roll);
 
   // Convert that into a world-space transform by taking the transpose of the inverse...
-  const glm::mat4 invOrientation = glm::transpose(glm::inverse(orientation));
+  const glm::dmat4 invOrientation = glm::transpose(glm::inverse(orientation));
 
-  up = (glm::vec3)(invOrientation * glm::vec4(0,1,0,0));
-  forward = (glm::vec3)(invOrientation * glm::vec4(0,0,-1,0));
-  right = (glm::vec3)(invOrientation * glm::vec4(1,0,0,0));
+  // Define new local axes...
+  up = (glm::dvec3)(invOrientation * glm::dvec4(0,1,0,0));
+  forward = (glm::dvec3)(invOrientation * glm::dvec4(0,0,-1,0));
+  right = (glm::dvec3)(invOrientation * glm::dvec4(1,0,0,0));
 
   // Move along the local XYZ axes...
-  const float elapsedSeconds = elapsedMS / 1000.0f;
-  const float velocity = elapsedSeconds * speed;
-  position += (glm::vec3)(velocity * glm::vec4(movement,0) * orientation);
-  movement = glm::vec3(0);
+  const double elapsedSeconds = elapsedMS / 1000.0f;
+  const double velocity = elapsedSeconds * speed;
+  position += (glm::dvec3)(velocity * glm::dvec4(movement,0) * orientation);
+  movement = glm::dvec3(0);
 
   // Compute new matrices...
   viewMatrix = orientation * glm::translate(-position);
